🎨: extract client error handling in new file
Extracting ResponseError and the logic to parse it into error.go to
reduce complexity in client.go.
